+++ CHANGELOG.md
@@ -7,6 +7,17 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 0.9.4 - Unreleased
 
 
+### Added
+* Added "resize" option to `stack_images_grid`
+* The `Coords` object now implements `__array__`.
+* The `about` parameter for `Polygon` transformations has been
+  expanded to include codes: "top,left", "top,right", etc...
+
+
+### Fixed
+* Fixed issue in copy methods of polygon and coords
+
+
 ## Version 0.9.3 - Released 2022-07-26
 
 
@@ -18,7 +29,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 * Support for decomposing projective transforms
 * Stack images now can do automatic casting between uint255 and float01
-* Added "resize" option to `stack_images_grid`
 
 
 ### Changed